pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "chii"
version = "0.8.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
# dependencies
pytz = "==2022.7"
loguru = "==0.6.0"
SQLAlchemy = { extras = ["mypy", "asyncio"], version = "==1.4.45" }
grpcio = "1.51.1"
grpcio-tools = "1.51.1"
libphpserialize = "0.0.8"
pymysql = "1.0.2"
pydantic = "1.10.4"

[tool.poetry.group.dev.dependencies]
sqlacodegen = "2.3.0"
# tests
coverage = { version = "==7.0.1", extras = ["toml"] }
pytest = "==7.2.0"
pytest-dotenv = "0.5.2"
# linter and formatter
pre-commit = "==2.21.0"
black = "==22.12.0"
flake8 = '4.0.1'
flake8-comprehensions = "==3.10.1"
flake8-bugbear = "==22.12.6"
flake8-broken-line = "==0.6.0"
flake8-pytest-style = "1.6.0"
flake8-tidy-imports = "==4.8.0"
mypy = "==0.991"
requests = "==2.28.1"
types-pytz = "2022.7.0.0"
pytest-github-actions-annotate-failures = "==0.1.8"
pytest-env = "==0.8.1"
redis = { extras = ["hiredis"], version = "==4.4.0" }
watchgod = "0.8.2"
types-protobuf = "^4.21.0.2"
types-redis = "^4.3.21.7"

[tool.coverage.run]
concurrency = ['thread', 'greenlet']

[tool.coverage.report]
exclude_lines = [
  'pragma: no cover',
  'if TYPE_CHECKING:',
  'if __name__ == "__main__":',
  'raise NotImplementedError'
]

[tool.pytest.ini_options]
addopts = '-rav -W error::sqlalchemy.exc.RemovedIn20Warning'
env = ['D:TESTING=true', 'SQLALCHEMY_WARN_20=true']

[tool.mypy]
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true
plugins = ['sqlalchemy.ext.mypy.plugin', 'pydantic.mypy']

[tool.black]
target_version = ['py38']
